requirements.txt
pycryptodome
keyring
pkcs7
requests